Login and Register implemented

+++ server/services/beerServices.js
@@ -1,6 +1,6 @@
 const Beer = require("../models/Beer");
 
-async function getAllBeersType(type) {
+async function getAllBeersType(beerType) {
   return await Beer.find({ type: beerType });
 }
 
@@ -15,13 +15,17 @@ async function getBeerById(id) {
 async function editBeer(id, data) {
   const beer = await Beer.findById(id);
 
-  (beer.name = data.name),
-    (beer.material = data.material),
-    (beer.country = data.country),
-    (beer.price = data.price),
-    (beer.imgUrl = data.imgUrl),
-    (beer.description = data.description),
-    (beer.type = data.type);
+  if (!beer) {
+    throw new Error("Beer not found");
+  }
+
+  beer.name = data.name;
+  beer.material = data.material;
+  beer.country = data.country;
+  beer.price = data.price;
+  beer.imgUrl = data.imgUrl;
+  beer.description = data.description;
+  beer.type = data.type;
 
   await beer.save();
 }
@@ -31,6 +35,11 @@ async function createBeer(data) {
 }
 
 async function delById(id) {
+  const beer = await Beer.findById(id);
+
+  if (!beer) {
+    throw new Error("Beer not found");
+  }
   await Beer.findByIdAndDelete(id);
 }
